refactor:NOT FOUND EXCEPTION 추상클래스화
