+++ src/main/java/jeje/work/aeatbe/exception/ArticleLikeNotFoundException.java
@@ -1,6 +1,6 @@
 package jeje.work.aeatbe.exception;
 
-public class ArticleLikeNotFoundException extends RuntimeException{
+public class ArticleLikeNotFoundException extends NotFoundException{
     public ArticleLikeNotFoundException(String message){
         super(message);
     }
