+++ src/main/java/jeje/work/aeatbe/exception/ColumnNotFoundException.java
@@ -1,6 +1,6 @@
 package jeje.work.aeatbe.exception;
 
-public class ColumnNotFoundException extends RuntimeException{
+public class ColumnNotFoundException extends NotFoundException{
     public ColumnNotFoundException(String message) {
         super(message);
     }
