+++ src/main/java/jeje/work/aeatbe/exception/NotFoundException.java
@@ -1,0 +1,7 @@
+package jeje.work.aeatbe.exception;
+
+public abstract class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
