+++ src/main/java/jeje/work/aeatbe/exception/UserNotFoundException.java
@@ -1,6 +1,6 @@
 package jeje.work.aeatbe.exception;
 
-public class UserNotFoundException extends RuntimeException{
+public class UserNotFoundException extends NotFoundException{
     public UserNotFoundException(String message) {
         super(message);
     }
